Change move number to u8.

+++ src/constants.rs
@@ -1,6 +1,6 @@
 use std::time::Duration;
 
-pub const MAX_MOVES_IN_GAME: usize = 102;
+pub const MAX_MOVES_IN_GAME: u8 = 102;
 pub const DEFAULT_TIME_LIMIT: Duration = Duration::from_secs(30);
 pub const TIME_MARGIN: Duration = Duration::from_millis(200);
